TheAgenticLoop/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==2.2.1
pydantic>=2.0.0

# Redis
redis>=5.0.0
fakeredis>=2.20.0

# AI Providers
groq>=0.4.0
cohere>=5.0.0
langchain-google-genai>=2.0.0

# Utils
python-dotenv>=1.0.0
bcrypt==4.2.0
